extract renderer
